multi raft logic encap, reply bug rectified
